vridge_back/requirements.txt
# Railway deploy trigger: 2025-07-31 15:30
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
dj-database-url==2.1.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0
psutil==5.9.6
boto3==1.29.7
django-storages==1.14.2
google-api-python-client==2.108.0
sendgrid==6.11.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.2
daphne==4.0.0

# AI/ML APIs
openai==1.3.0
google-generativeai==0.3.0

# PDF generation
reportlab==4.0.4

# Video Analysis
twelvelabs==0.2.0